requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.5.0
pandas>=2.0.0
jax[cpu]>=0.4.13
jaxopt>=0.7
pydantic>=2.4.0
impedance>=1.7.0

# LLM API
openai>=1.3.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# CLI
typer>=0.9.0
rich>=13.6.0

# Async/Concurrent Processing
aiofiles>=23.2.1
asyncio>=3.4.3

# Excel Support
openpyxl>=3.1.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=23.9.1